requirements.txt
asgiref==3.3.1
autopep8==1.5.3
cursor==1.3.4
Django==2.1.1
pycodestyle==2.6.0
pymemcache==3.4.1
PyMySQL==1.0.2
python-memcached==1.59
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1




